# backend/src/main/java/hu/bme/itsec/simnyi/backend/model/PasswordDTO.java
package hu.bme.itsec.simnyi.backend.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotBlank;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class PasswordDTO   {

  @NotBlank
  private String password;

}